feat(backend): store compression dictionary ids in database, also still read data that has no dictionary id stored (#5311)
first part of #4769 as mentioned in
https://github.com/loculus-project/loculus/issues/4769#issuecomment-3435703053

This adds a `compression_dictionaries` table to the DB. It does not
migrate existing data. The backend will be changed to write only entries
in the new format (with the dict id) and to read both formats (with and
without the dict it - preserving the current logic).

Next step: https://github.com/loculus-project/loculus/pull/5383

## Flyway migration logic changes

I changed how the `Flyway` object is created: The backend no longer
instantiates it itself, it leaves the creation to the autoconfiguration
that comes with its Spring boot integration. Mostly the consequences
are:
* We have to call `migrate` somewhere else, but that's not really an
issue.
* We can configure Flyway through Spring properties (e.g. in
`application.properties`)
* We can use Spring beans in Java based migration scripts (that's why I
did that - the autoconfig correctly wires Spring and Flyway together so
that it works, I didn't want to go through that myself)

## Compression Changes
The `CompressedSequence` class is updated to include the key of a
compression dictionary in the compression dictionary table (this is
optional as older existing entries will not have this key). This is
added to the JSONB object in the database.

New unalignedSequences in originalData are compressed using a dictionary
created from a concatenated string of all segments. Compression of
processed sequences happens as before.

Decompression now uses the compression dict Id to decompress sequences
and takes a compression dict object as a fallback for existing entries
that do not have an Id. This object is always calculated based on the
segment for existing entries.

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: GitHub Action <action@github.com>