fix(integration-tests): don't create groups with duplicate names to avoid flakes (#5459)
Fixes #5457

We used to create duplicate groups when the readonly group was not
visible in dependent tests. This caused those tests to then fail as the
actual id was different and no sequence existed.

In this PR I refactor functions related to getting ids of existing
groups to make it more explicit if duplicates are to be created (usually
not).

Also make a few tests more robust and extend timeouts where they were
hit (e.g. here:
https://github.com/loculus-project/loculus/actions/runs/19461908991/job/55687839118)

Add .nvmrc for integration-tests