fix(ci): Switch website image waiting from wait-on-check-action to polling docker for image existence
Now that we build the website image in a multi-job way, the wait-on-check-action no longer works for waiting on website image to be ready.

Instead, we can just poll for the website image to be available on ghcr.

This is actually less brittle in that it doesn't rely on a particular action's name.

Fixes this issue: https://github.com/loculus-project/loculus/actions/runs/19707963135/job/56460456942

```
The requested check was never run against this ref, exiting...
```