fix(deployment): add exponential backoff retry for helm repo add (#5590)
Fixes #5588

We are seeing intermittent timeouts when adding the mittwald helm repo
(at least twice today, I noticed it on just a few runs and I remember
seeing it last week as well).

This PR adds an exponential backoff retry mechanism to the `helm repo
add` command in `deploy.py`.
It will retry up to 5 times (6 attempts total) with a delay starting at
5 seconds and doubling up to 60 seconds.

🚀 Preview: Add `preview` label to enable