feat(website): edit page - do not include fasta description in metadata field `fastaIds` (#5629)
resolves https://github.com/loculus-project/loculus/issues/5627

The `fastaIds` object should contain a space separated list of fasta
IDs. The website was constructing `fastaIds` for submissions via the
edit page by concatenating full fasta headers of the form `>fastaId
description` (and removing `>`) - but this lead to the backend thinking
the description was another fastaId.

Also adds a description to multiple fasta files in the integration tests
to ensure in future such an issue cannot be missed

### Screenshot

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://fastaheader.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>
Co-authored-by: Fabian Engelniederhammer <92720311+fengelniederhammer@users.noreply.github.com>