fix(prepro): no-alignment, no sequences option (#5654)
resolves https://github.com/loculus-project/loculus/issues/5653

### Screenshot
The backend returns `"unalignedNucleotideSequences": {}` for the case
where we run Loculus without any sequences, as prepro now tries to
assign the segment using the fasta header or the dictioanry keys (as
there is no nextclade dataset prepro defaults to the behavior previously
seen in the backend) prepro fails with an internal server error as the
dictionary does not have any keys.

### PR Checklist
- [ ] All necessary documentation has been adapted.
- [ ] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
- Tested on WISE instance

🚀 Preview: Add `preview` label to enable