fix(deposition): correct version type (#5651)
part of https://github.com/loculus-project/loculus/issues/5646

Version is an int in the db (see
https://github.com/loculus-project/loculus/blob/main/ena-submission/flyway/sql/V1__Initial_Schema.sql#L3)-
this was incorrectly annotated in the code (python doesn't care) - this
is a step towards improving typing in python (we will later introduce
pydantic or typeddicts which will validate all entries in the db are the
correct specified types.

### PR Checklist
- ~All necessary documentation has been adapted.~
- [x] The implemented feature is covered by appropriate, automated
tests.
- ~Any manual testing that has been done is documented (i.e. what
exactly was tested?)~ not required as integration tests will submit to
ENA dev

🚀 Preview: Add `preview` label to enable