feat(backend): Expose a subset of group info to users that are not logged in (#5647)
resolves https://github.com/loculus-project/loculus/issues/5615 (first
part)

### Background
We want to expose some group information when requests to
/groups/{groupID} are made to the backend without authentication:
- Group details with the emailContact and user fields set to null should
be returned when a GET request is made to the backend at
groups/{groupId}.
- Full set of information should be returned by a GET request to the
same endpoint by an authenticated user.

### Overview of changes:
- "/groups/*" added to "getEndpointsThatArePublic"
- "getDetailsOfGroup()" checks if user is logged in and decides what to
return
- "GroupDetails" data class 'users' field is now optional
- "Group" data class 'emailContact' field is now optional
- To create a "Group" with redacted email, use
"GroupEntity.toGroup(redactEmail = true)"

### What I tested manually
`curl` GET requests to the endpoint behave as expected: contactEmail is
included when using an authentication token, and set to N/A when there
is no token provided. I also checked the swagger-ui to verify that the
documentation there is updated and reflects the new behaviour.

### Open issues & questions:
- Do we want to do something similar for the "/groups" endpoint? It
could return a list of all Groups with the contact info redacted.

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)

🚀 Preview: https://expose-group-info.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>
Co-authored-by: Anna (Anya) Parker <50943381+anna-parker@users.noreply.github.com>