refactor(deposition): jsonb columns are lists or dicts in python, move json.dumps to db upload functions (#5659)
part of https://github.com/loculus-project/loculus/issues/5646

Keeping inline with the Loculus backend the ena-deposition DB has jsonb
fields. These fields correspond to lists or dictionaries in the code but
must be "dumped" into a string object when storing in the database. This
was done inconsistently in the code.

Now the dumping is only performed when actually writing to the DB. This
allows me to refactor the entry objects to have the types we use in the
code - making the setup clearer.

### PR Checklist
- ~All necessary documentation has been adapted.~
- [x] The implemented feature is covered by appropriate, automated
tests.
- [ ] Any manual testing that has been done is documented (i.e. what
exactly was tested?) -> the integration tests should take away the need
for manual testing but I can show @maverbiest how to ssh into the
database on a preview so you can look at what the entries look like
there :-)

🚀 Preview: Add `preview` label to enable