feat(deposition): add unit tests for the cronjob that uses the backend's get-released-data format (#5226)
This will help us to ensure we don't break anything if we ever refactor
the cronjob.

The cronjob calls get-released-data and then uses that data to create a
json with all data that we need to submit (this might change in future)
- this means it partially transforms data. It also checks if data has
already been submitted by us.

Previously we had no tests to cover this transformation. Or the database
check. This unittest mocks slack calls, db checks and backend requests,
however it checks that the JSON is formatted correctly and that we will
send a slack notification with the correct text.

### Screenshot

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- ~Any manual testing that has been done is documented (i.e. what
exactly was tested?)~ this does not touch any code but only adds tests

@maverbiest I know this is a lot to review - if you dont feel
comfortable approving that is fine - but it would be cool if you could
have a look at the code and ask me if you have any follow up questions
:-)

🚀 Preview: Add `preview` label to enable

---------

Co-authored-by: maverbiest <59956362+maverbiest@users.noreply.github.com>
Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>