feat(deposition): retry if biosample, bioproject or run_ref are now visible (#5267)
Currently when submitting assemblies with associated raw reads we have
to wait for the biosample, bioproject and raw reads to become public or
else the submission will fail. We can also submit the sequences before
they are public but as we don't have an automatic re-try option we need
to manually update the DB to have the pipeline retry submission.

This takes care of retrying for us. It **only retries specific cases**,
i.e. when the biosample, bioproject or raw reads associated with an
assembly were not public. It also continues to send slack notifications
about this problem. Also to not overload ENA the time between retries is
configurable and I have set it to 4hours by default.

### Screenshot

### PR Checklist
- [x] All necessary documentation has been adapted.
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
Created a group 2, updated DB test entries to have groupId 2, after
submission succeeded. Set the assembly from state WAITING to HAS_ERRORS
in the db - confirmed that the entry was almost immediately retried.
Reset the state again and then confirmed the pod waited the specified
time (set to 4min)
<img width="2404" height="872" alt="image"
src="https://github.com/user-attachments/assets/6e1dadff-1d9b-436c-93e1-6f0b9750f7f9"
/>

🚀 Preview: https://retry-ena-submissions.loculus.org

---------

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>