feat!(config, deposition): extend ENA deposition to multi-path organisms (#5668)
resolves https://github.com/loculus-project/loculus/issues/5512

Changes the ena-deposition config, to now have the same structure as the
referenceSequences - see below. This PR also updates the templates to
pass the segments to the ena deposition pod in the config and use when
submitting.

This PR makes little changes to the core ena deposition code. It defines
"ENA organisms" that the pod uploads to and creates bioprojects for.
This is used to submit subtypes to their subtaxon instead of the taxon
of the superorganism, but it still uploads the results of the submission
to the correct "loculus Organism".

### Config changes
This is only required if you are using the ENA deposition feature:
```
enaDeposition:
  configFile:
    taxon_id: 3052518
    scientific_name: "Orthonairovirus haemorrhagiae"
    molecule_type: "genomic RNA"
```
becomes:
```
enaDeposition:
  singleReference:
    configFile:
      taxon_id: 3052518
      scientific_name: "Orthonairovirus haemorrhagiae"
      molecule_type: "genomic RNA"
```
for the multi-path case `singleRefence` is replaced with the
`subOrganism` name - and this is used for ENA submissions.


### PR Checklist
- [x] Change config as suggested in
https://github.com/loculus-project/loculus/issues/5512
- [x] ingest data from super-organism
- [x] No longer assume all single segmented organisms have the segment
name main, get this from the config
- [x] submit results to correct loculusOrganism
- [x] The implemented feature is covered by appropriate, automated
tests.
- [x] Any manual testing that has been done is documented (i.e. what
exactly was tested?)
Uploaded ebola and EV sequences, triggered cronjob and checked this
works: https://loculus.slack.com/archives/C07FP0BGHDZ/p1765617124903149.
Then uploaded the sequences to confirm that the upload is successful
with the changes in this PR:
<img width="2402" height="114" alt="image"
src="https://github.com/user-attachments/assets/4cadf9e4-e375-4c1b-be3a-a2d45eb02337"
/>
Sequences in question:
https://multi-path-ena.loculus.org/seq/LOC_001EDWZ.1
https://multi-path-ena.loculus.org/seq/LOC_001EER8.1 (note bioproject
and biosample were uploaded correctly for both organisms!)

🚀 Preview: https://multi-path-ena.loculus.org